Rename triforce_lib
