+++ triforce/custom_integrations/Zelda-NES/0_67s.state
