+++ triforce/custom_integrations/Zelda-NES/0_77.state
