+++ triforce/custom_integrations/Zelda-NES/0_78w.state
