+++ triforce/custom_integrations/Zelda-NES/1_23s.state
