+++ triforce/custom_integrations/Zelda-NES/1_35s.state
