+++ triforce/custom_integrations/Zelda-NES/1_36w.state
