+++ triforce/custom_integrations/Zelda-NES/1_42s.state
