+++ triforce/custom_integrations/Zelda-NES/1_43n.state
