+++ triforce/custom_integrations/Zelda-NES/1_44w.state
