+++ triforce/custom_integrations/Zelda-NES/1_45w.state
