+++ triforce/custom_integrations/Zelda-NES/1_52e.state
