+++ triforce/custom_integrations/Zelda-NES/1_72e.state
