+++ triforce/custom_integrations/Zelda-NES/1_73s.state
