+++ triforce/custom_integrations/Zelda-NES/1_74w.state
