+++ triforce/custom_integrations/Zelda-NES/1_83e.state
