+++ triforce/custom_integrations/Zelda-NES/scenario.json
