+++ triforce/custom_integrations/Zelda-NES/start.state
