+++ triforce/end_conditions.py
