+++ triforce/zelda_game_data.txt
